Divide Launcher into Converter and Benchmarker
