+++ netspresso/benchmarker/__init__.py
@@ -1,0 +1,4 @@
+from .benchmarker import Benchmarker
+
+
+__all__ = ["Benchmarker"]
