--- netspresso/clients/compressor/enums/__init__.py
@@ -1,34 +1,0 @@
-from .compression import (
-    CompressionMethod,
-    GroupPolicy,
-    LayerNorm,
-    Policy,
-    RecommendationMethod,
-    compression_literal,
-    recommendation_literal,
-    policy_literal,
-    grouppolicy_literal,
-    layernorm_literal,
-)
-from .model import Extension, Framework, OriginFrom, Task, task_literal, framework_literal, extension_literal, originfrom_literal
-
-__all__ = [
-    "CompressionMethod",
-    "RecommendationMethod",
-    "Policy",
-    "GroupPolicy",
-    "LayerNorm",
-    "Task",
-    "Framework",
-    "Extension",
-    "OriginFrom",
-    "compression_literal",
-    "recommendation_literal",
-    "policy_literal",
-    "grouppolicy_literal",
-    "layernorm_literal",
-    "task_literal",
-    "framework_literal",
-    "extension_literal",
-    "originfrom_literal",
-]
