+++ netspresso/enums/__init__.py
@@ -1,4 +1,31 @@
 from .credit import ServiceCredit
 from .metadata import TaskType, Status
+from .compression import CompressionMethod, RecommendationMethod, Policy, GroupPolicy, LayerNorm
+from .model import Task, Framework, Framework, Extension, OriginFrom, DataType
+from .device import DeviceName, SoftwareVersion, HardwareType, TaskStatus
+from .module import Module
+from netspresso.clients.compressor.schemas.compression import Options
 
-__all__ = ["ServiceCredit", "TaskType", "Status"]
+
+__all__ = [
+    "ServiceCredit",
+    "TaskType",
+    "Status",
+    "CompressionMethod",
+    "RecommendationMethod",
+    "Policy",
+    "GroupPolicy",
+    "LayerNorm",
+    "Task",
+    "Framework",
+    "Framework",
+    "Extension",
+    "OriginFrom",
+    "DataType",
+    "DeviceName",
+    "SoftwareVersion",
+    "HardwareType",
+    "TaskStatus",
+    "Module",
+    "Options",
+]
