+++ netspresso/enums/compression.py
