+++ netspresso/enums/device.py
@@ -1,0 +1,100 @@
+from enum import Enum
+from typing import Literal
+
+
+class DeviceName(str, Enum):
+    RASPBERRY_PI_4B = "RaspberryPi4B"
+    RASPBERRY_PI_3B_PLUS = "RaspberryPi3BPlus"
+    RASPBERRY_PI_ZERO_W = "RaspberryPi-ZeroW"
+    RASPBERRY_PI_ZERO_2W = "RaspberryPi-Zero2W"
+    RENESAS_RZ_V2L = "rzv2l_avnet"
+    RENESAS_RZ_V2M = "rzv2m"
+    JETSON_NANO = "Jetson-Nano"
+    JETSON_TX2 = "Jetson-Tx2"
+    JETSON_XAVIER = "Jetson-Xavier"
+    JETSON_NX = "Jetson-Nx"
+    JETSON_AGX_ORIN = "Jetson-AGX-Orin"
+    AWS_T4 = "AWS-T4"
+    INTEL_XEON_W_2233 = "Intel-Xeon"
+    ALIF_ENSEMBLE_E7_DEVKIT_GEN2 = "Ensemble-E7-DevKit-Gen2"
+    RENESAS_RA8D1 = "Renesas-RA8D1"
+
+    @classmethod
+    def create_literal(cls):
+        return Literal[
+            "RaspberryPi4B",
+            "RaspberryPi3BPlus",
+            "RaspberryPi-ZeroW",
+            "RaspberryPi-Zero2W",
+            "rzv2l_avnet",
+            "rzv2m",
+            "Jetson-Nano",
+            "Jetson-Tx2",
+            "Jetson-Xavier",
+            "Jetson-Nx",
+            "Jetson-AGX-Orin",
+            "AWS-T4",
+            "Intel-Xeon",
+            "Ensemble-E7-DevKit-Gen2",
+            "Renesas-RA8D1",
+        ]
+    
+    JETSON_DEVICES = [
+        JETSON_NANO,
+        JETSON_TX2,
+        JETSON_XAVIER,
+        JETSON_NX,
+        JETSON_AGX_ORIN,
+    ]
+    RASPBERRY_PI_DEVICES = [
+        RASPBERRY_PI_4B,
+        RASPBERRY_PI_3B_PLUS,
+        RASPBERRY_PI_ZERO_W,
+        RASPBERRY_PI_ZERO_2W,
+    ]
+    RENESAS_DEVICES = [RENESAS_RZ_V2L, RENESAS_RZ_V2M]
+    NVIDIA_GRAPHIC_CARDS = [AWS_T4]
+    INTEL_DEVICES = [INTEL_XEON_W_2233]
+    AVAILABLE_INT8_DEVICES = [
+        ALIF_ENSEMBLE_E7_DEVKIT_GEN2,
+        RENESAS_RA8D1,
+        RASPBERRY_PI_4B,
+        RASPBERRY_PI_3B_PLUS,
+        RASPBERRY_PI_ZERO_W,
+        RASPBERRY_PI_ZERO_2W,
+    ]
+    ONLY_INT8_DEVICES = [ALIF_ENSEMBLE_E7_DEVKIT_GEN2, RENESAS_RA8D1]
+
+
+class SoftwareVersion(str, Enum):
+    JETPACK_4_4_1 = "4.4.1-b50"
+    JETPACK_4_6 = "4.6-b199"
+    JETPACK_5_0_1 = "5.0.1-b118"
+    JETPACK_5_0_2 = "5.0.2-b231"
+
+    @classmethod
+    def create_literal(cls):
+        return Literal["4.4.1-b50", "4.6-b199", "5.0.1-b118", "5.0.2-b231"]
+
+
+class HardwareType(str, Enum):
+    HELIUM = "helium"
+
+    @classmethod
+    def create_literal(cls):
+        return Literal["helium"]
+
+
+class TaskStatus(str, Enum):
+    IN_QUEUE = "IN_QUEUE"
+    IN_PROGRESS = "IN_PROGRESS"
+    FINISHED = "FINISHED"
+    ERROR = "ERROR"
+    USER_CANCEL = "USER_CANCEL"
+
+    @classmethod
+    def create_literal(cls):
+        return Literal["IN_QUEUE", "IN_PROGRESS", "FINISHED", "ERROR", "USER_CANCEL"]
+
+
+device_name_literal = DeviceName.create_literal()
