+++ netspresso/enums/module.py
@@ -1,0 +1,11 @@
+from enum import Enum
+from typing import Literal
+
+
+class Module(str, Enum):
+    CONVERT = "CONVERT"
+    BENCHMARK = "BENCHMARK"
+
+    @classmethod
+    def create_literal(cls):
+        return Literal["CONVERT", "BENCHMARK"]
